Delegate error handling/thread management of BackgroundOpsManager to separate classes
